requirements.txt
appdirs==1.4.3
attrs==19.3.0
autopep8==1.4.4
black==19.10b0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pathspec==0.6.0
pycodestyle==2.5.0
PyJWT==1.7.1
pyodbc==4.0.27
regex==2019.12.20
six==1.13.0
SQLAlchemy==1.3.12
toml==0.10.0
typed-ast==1.4.0
Werkzeug==0.16.0
yapf==0.29.0
